Backup feature.
- Added backup agent to remotely save SharedPreferences file
- Added privacy policy, to conforms to the Android Backup Service terms of use
"4.10 You agree that in your use of the Service, you and your Applications will protect the privacy and legal rights of users. You must provide legally adequate privacy notice and protection for users whose data your Applications back up to the Service. Further, your Application may only use that information for the limited purpose of backing up the data to the Service unless the user has given you permission for further use. If the user has not given you permission to back up information to the Service, you may not transmit such information to the Service."
https://developer.android.com/google/backup/terms.html